els-ASAP auto commit file/tag qemu-guest-agent-2.5.0-3.el7.src.rpm

+++ src/roms/SLOF/lib/libc/include/stddef.h
@@ -16,7 +16,8 @@
 
 #define NULL ((void *)0)
 
-typedef unsigned long size_t;
+
+typedef unsigned int size_t;
 
 
 #endif
